alternate phrasing
caveat: I haven't run the tests because I'm feeling a bit out of sorts. but my goal here was to make the code easier to read. using `case` and `when` over `if` and `elsif` is a definite thing; it's more Ruby-ish and it's almost always the way to go. but it's more of a personal judgement call when it comes to using

`adjectives = (case/when)...`

versus

```ruby
case foo
when bar
  adjectives = "one thing"
when baz
  adjectives = "another thing"
end
```

that is to say, when it comes to not repeating the `adjectives =` part. in this case though I think it's the right move.

I don't know what's going on with this last part:

```ruby
hash = Hash[adjectives.each_with_index.map{|word, idx| [idx, word]}]
hash.to_json
```

my hunch is to say "rewrite it," but I'd like to know more about it first.